Deleted extra files & spacing as well as updated comments
