+++ .idea/sqldialects.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/test/resources/cleanDB.sql" dialect="MySQL" />
+    <file url="file://$PROJECT_DIR$/src/test/resources/cleandb.sql" dialect="MySQL" />
   </component>
 </project>
