--- src/main/java/edu/matc/restdemo/HelloWorld.java
@@ -1,19 +1,0 @@
-package edu.matc.restdemo;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Response;
-
-@Path("/greetings")
-public class HelloWorld {
-    // The Java method will process HTTP GET requests
-    @GET
-    // The Java method will produce content identified by the MIME Media type "text/plain"
-    @Produces("text/plain")
-    public Response getMessage() {
-        // Return a simple message
-        String output = "Hello World";
-        return Response.status(200).entity(output).build();
-    }
-}
