+++ src/main/java/edu/matc/restdemo/Regions.java
@@ -14,25 +14,34 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
 import java.util.List;
 
+/**
+ * The type Regions.
+ *
+ * @author jbjohnson2
+ * @author JBostroem
+ * @author OscarJohnson6
+ */
 @Path("/regions")
 public class Regions {
     private final Logger logger = LogManager.getLogger(this.getClass());
+    String regionsString = "Error making json file";
 
+    /**
+     * Gets regions.
+     *
+     * @return the regions
+     */
     @GET
     @Produces("application/json")
     public Response getRegions() {
         ObjectMapper objectMapper = new ObjectMapper();
         GenericDao<Region> regionDao = new GenericDao<>(Region.class);
         List<Region> regions = regionDao.getAll();
-        String regionsString = "Error making json file";
 
         try {
             regionsString = objectMapper.writeValueAsString(regions);
 
-            logger.info(regionsString);
-            logger.info(Response.status(200).entity(regionsString).build());
             return Response.status(200).entity(regionsString).build();
-
         } catch (JsonProcessingException jsonProcessingException) {
             logger.error("Error making json file", jsonProcessingException);
         }
@@ -40,6 +49,12 @@ public class Regions {
         return Response.status(500).entity(regionsString).build();
     }
 
+    /**
+     * Gets region by id.
+     *
+     * @param id the id
+     * @return the region by id
+     */
     @GET
     @Path("/{id}")
     @Produces("application/json")
@@ -48,19 +63,14 @@ public class Regions {
         GenericDao<Region> regionDao = new GenericDao<>(Region.class);
         Region region = regionDao.getById(id);
 
-        String regionString = "Error making json file";
-
         try {
-            regionString = objectMapper.writeValueAsString(region);
+            regionsString = objectMapper.writeValueAsString(region);
 
-            logger.info(regionString);
-            logger.info(Response.status(200).entity(regionString).build());
-            return Response.status(200).entity(regionString).build();
-
+            return Response.status(200).entity(regionsString).build();
         } catch (JsonProcessingException jsonProcessingException) {
             logger.error("Error making json file", jsonProcessingException);
         }
 
-        return Response.status(500).entity(regionString).build();
+        return Response.status(500).entity(regionsString).build();
     }
 }
