+++ src/main/java/edu/matc/restdemo/Types.java
@@ -11,32 +11,47 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.Response;
 import java.util.List;
 
+/**
+ * The type Types.
+ *
+ * @author jbjohnson2
+ * @author JBostroem
+ * @author OscarJohnson6
+ */
 @Path("/types")
 public class Types {
     private final Logger logger = LogManager.getLogger(this.getClass());
+    String typeString = "Error making json file";
 
+    /**
+     * Gets types.
+     *
+     * @return the types
+     */
     @GET
     @Produces("application/json")
     public Response getTypes() {
         ObjectMapper objectMapper = new ObjectMapper();
         GenericDao<Type> typeDao = new GenericDao<>(Type.class);
         List<Type> types = typeDao.getAll();
-        String typesString = "Error making json file";
 
         try {
-            typesString = objectMapper.writeValueAsString(types);
+            typeString = objectMapper.writeValueAsString(types);
 
-            logger.info(typesString);
-            logger.info(Response.status(200).entity(typesString).build());
-
-            return Response.status(200).entity(typesString).build();
+            return Response.status(200).entity(typeString).build();
         } catch (JsonProcessingException jsonProcessingException) {
             logger.error("Error making json file", jsonProcessingException);
         }
 
-        return Response.status(500).entity(typesString).build();
+        return Response.status(500).entity(typeString).build();
     }
 
+    /**
+     * Gets type by id.
+     *
+     * @param id the id
+     * @return the type by id
+     */
     @GET
     @Path("/{id}")
     @Produces("application/json")
@@ -44,13 +59,9 @@ public class Types {
         ObjectMapper objectMapper = new ObjectMapper();
         GenericDao<Type> typeDao = new GenericDao<>(Type.class);
         Type type = typeDao.getById(id);
-        String typeString = "Error making json file";
 
         try {
             typeString = objectMapper.writeValueAsString(type);
-
-            logger.info(typeString);
-            logger.info(Response.status(200).entity(typeString).build());
 
             return Response.status(200).entity(typeString).build();
         } catch (JsonProcessingException jsonProcessingException) {
